Refactor to run this on a single apply
I've moved all of the files out of the components directory and moved
them into the root.

I've created a single provider file and updated it to use the S3 bucket
in the management account.

The original bucket in the provider files doesn't seem to exist so I had
to recreate the state files.

The terraform files have been copied and pasted verbatim. This PR is
about getting it to deploy, not tidying everything up so I'm not going
to change any names or use da-terraform-modules at the moment.  We can
do this later.